Treesame-Commit-Id: c8e0a70c867229110dad59339242445bb543f940

+++ db/migrate/20120224194848_remove_unused_notifications.rb
@@ -2,6 +2,7 @@ class RemoveUnusedNotifications < ActiveRecord::Migration
   tag :postdeploy
 
   def self.up
+    return unless Shard.current.default?
     Notification.update_all({:category => 'Other'}, :category => 'Message')
     Notification.delete_all(:category => ['Files', 'Student Message'])
   end
